src/app/core/modules/header/header.component.ts
import {ChangeDetectionStrategy, Component} from '@angular/core';
import {Store} from '@ngrx/store';
import {AppState} from '../../store';
import * as AuthActions from '../auth/store/auth.actions';

@Component({
  selector: 'app-header',
  templateUrl: './header.component.html',
  styleUrls: ['./header.component.scss'],
  changeDetection: ChangeDetectionStrategy.OnPush,
})
export class HeaderComponent {

  constructor(private store: Store<AppState>) {
  }

  onSignOut(): void {
    this.store.dispatch(AuthActions.PageSignOutAction());
  }
}
